rename llama_stack modules and paths to ogx
