+++ docker-compose.yaml
@@ -3,7 +3,7 @@ services:
   llama-stack:
     build:
       context: .
-      dockerfile: deploy/llama-stack/test.containerfile
+      dockerfile: deploy/ogx/test.containerfile
     platform: linux/amd64
     container_name: llama-stack
     ports:
@@ -18,8 +18,8 @@ services:
       # from lightspeed-stack.yaml and this mount is ignored
       - ./run.yaml:/opt/app-root/run.yaml:z
       # Host copies so `docker compose up` picks up script changes without rebuilding OGX
-      - ./scripts/llama-stack-entrypoint.sh:/opt/app-root/enrich-entrypoint.sh:ro,z
-      - ./src/llama_stack_configuration.py:/opt/app-root/llama_stack_configuration.py:ro,z
+      - ./scripts/ogx-entrypoint.sh:/opt/app-root/enrich-entrypoint.sh:ro,z
+      - ./src/ogx_configuration.py:/opt/app-root/ogx_configuration.py:ro,z
       - ./src/data:/opt/app-root/data:ro,z
       - ${GCP_KEYS_PATH:-./tmp/.gcp-keys-dummy}:/opt/app-root/.gcp-keys:ro
       - ./lightspeed-stack.yaml:/opt/app-root/lightspeed-stack.yaml:ro,z
