+++ src/client/ogx.py
@@ -15,16 +15,16 @@ from ogx_client import APIConnectionError, APIStatusError, AsyncOgxClient
 import constants
 from authorization.azure_token_manager import AzureEntraIDManager
 from configuration import configuration
-from llama_stack_configuration import (
+from log import get_logger, setup_logging
+from models.api.responses.error import ServiceUnavailableResponse
+from models.config import OgxConfiguration
+from ogx_configuration import (
     YamlDumper,
     enrich_azure_entra_id_inference,
     enrich_byok_rag,
     enrich_solr,
     synthesize_to_file,
 )
-from log import get_logger, setup_logging
-from models.api.responses.error import ServiceUnavailableResponse
-from models.config import OgxConfiguration
 from utils.model_list import parse_model_list_response
 from utils.types import Singleton
