+++ src/constants.py
@@ -202,7 +202,7 @@ CACHE_TYPE_POSTGRES: Final[str] = "postgres"
 CACHE_TYPE_NOOP: Final[str] = "noop"
 
 # BYOK RAG
-# Backends that have enrichment support in llama_stack_configuration.py
+# Backends that have enrichment support in ogx_configuration.py
 SUPPORTED_RAG_BACKENDS: Final[frozenset[str]] = frozenset({"faiss", "pgvector"})
 
 # Default RAG backend for bring-your-own-knowledge RAG configurations
