+++ src/data/__init__.py
@@ -2,7 +2,7 @@
 
 Currently holds ``default_run.yaml``, the built-in baseline OGX
 configuration used by unified-mode synthesis (see
-``llama_stack_configuration.load_default_baseline``). Making this directory a
+``ogx_configuration.load_default_baseline``). Making this directory a
 package ensures the data file is included in built wheels and resolvable both
 in editable installs and from site-packages.
 """
