--- src/pydantic_ai_lightspeed/llamastack/__init__.py
@@ -1,6 +1,0 @@
-"""Pydantic AI provider for OGX."""
-
-from pydantic_ai_lightspeed.llamastack._model import OgxResponsesModel
-from pydantic_ai_lightspeed.llamastack._provider import OgxProvider
-
-__all__ = ["OgxProvider", "OgxResponsesModel"]
