+++ src/pydantic_ai_lightspeed/ogx/README.md
@@ -1,4 +1,4 @@
-# List of source files stored in `src/pydantic_ai_lightspeed/llamastack` directory
+# List of source files stored in `src/pydantic_ai_lightspeed/ogx` directory
 
 ## [__init__.py](__init__.py)
