+++ src/pydantic_ai_lightspeed/ogx/__init__.py
@@ -1,0 +1,6 @@
+"""Pydantic AI provider for OGX."""
+
+from pydantic_ai_lightspeed.ogx._model import OgxResponsesModel
+from pydantic_ai_lightspeed.ogx._provider import OgxProvider
+
+__all__ = ["OgxProvider", "OgxResponsesModel"]
