+++ src/utils/pydantic_ai_helpers.py
@@ -23,7 +23,7 @@ from models.config import (
 )
 from pydantic_ai_lightspeed.capabilities import QuestionValidity
 from pydantic_ai_lightspeed.capabilities.redaction import PiiRedactionCapability
-from pydantic_ai_lightspeed.llamastack import OgxResponsesModel
+from pydantic_ai_lightspeed.ogx import OgxResponsesModel
 from utils.shields import get_shields_for_request
 
 _AGENT_SKILLS_PROVIDER_ID: Final[str] = "agent-skills"
