+++ tests/e2e-prow/rhoai/manifests/lightspeed/ogx-prow.yaml
@@ -157,7 +157,7 @@ spec:
           if [[ -f "$LIGHTSPEED_CONFIG" ]]; then
             echo "Enriching llama-stack config..."
             ENRICHMENT_FAILED=0
-            /opt/app-root/.venv/bin/python3 /opt/app-root/src/llama_stack_configuration.py \
+            /opt/app-root/.venv/bin/python3 /opt/app-root/src/ogx_configuration.py \
               -c "$LIGHTSPEED_CONFIG" \
               -i "$INPUT_CONFIG" \
               -o "$ENRICHED_CONFIG" 2>&1 || ENRICHMENT_FAILED=1
