+++ tests/unit/README.md
@@ -24,11 +24,11 @@ Unit tests for the degraded mode tracker.
 
 Unit tests for functions defined in src/lightspeed_stack.py.
 
-## [test_llama_stack_configuration.py](test_llama_stack_configuration.py)
+## [test_ogx_configuration.py](test_ogx_configuration.py)
 
-Unit tests for src/llama_stack_configuration.py.
+Unit tests for src/ogx_configuration.py.
 
-## [test_llama_stack_synthesize.py](test_llama_stack_synthesize.py)
+## [test_ogx_synthesize.py](test_ogx_synthesize.py)
 
 Unit tests for unified-mode OGX configuration synthesis (LCORE-2336).
