+++ tests/unit/models/config/test_ogx_configuration.py
