--- tests/unit/pydantic_ai_lightspeed/llamastack/README.md
@@ -1,18 +1,0 @@
-# List of source files stored in `tests/unit/pydantic_ai_lightspeed/llamastack` directory
-
-## [__init__.py](__init__.py)
-
-Unit tests for pydantic_ai_lightspeed.llamastack sub-package.
-
-## [test_model.py](test_model.py)
-
-Unit tests for pydantic_ai_lightspeed.llamastack._model module.
-
-## [test_provider.py](test_provider.py)
-
-Unit tests for pydantic_ai_lightspeed.llamastack._provider module.
-
-## [test_transport.py](test_transport.py)
-
-Unit tests for pydantic_ai_lightspeed.llamastack._transport module.
-
