--- tests/unit/pydantic_ai_lightspeed/llamastack/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for pydantic_ai_lightspeed.llamastack sub-package."""
