+++ tests/unit/pydantic_ai_lightspeed/ogx/README.md
@@ -1,0 +1,18 @@
+# List of source files stored in `tests/unit/pydantic_ai_lightspeed/ogx` directory
+
+## [__init__.py](__init__.py)
+
+Unit tests for pydantic_ai_lightspeed.ogx sub-package.
+
+## [test_model.py](test_model.py)
+
+Unit tests for pydantic_ai_lightspeed.ogx._model module.
+
+## [test_provider.py](test_provider.py)
+
+Unit tests for pydantic_ai_lightspeed.ogx._provider module.
+
+## [test_transport.py](test_transport.py)
+
+Unit tests for pydantic_ai_lightspeed.ogx._transport module.
+
