+++ tests/unit/pydantic_ai_lightspeed/ogx/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for pydantic_ai_lightspeed.ogx sub-package."""
