+++ tests/unit/pydantic_ai_lightspeed/ogx/test_provider.py
@@ -1,4 +1,4 @@
-"""Unit tests for pydantic_ai_lightspeed.llamastack._provider module."""
+"""Unit tests for pydantic_ai_lightspeed.ogx._provider module."""
 
 # pylint: disable=protected-access
 
@@ -9,11 +9,11 @@ from ogx_client import AsyncOgxClient
 from openai import AsyncOpenAI
 from pytest_mock import MockerFixture
 
-from pydantic_ai_lightspeed.llamastack._provider import (
+from pydantic_ai_lightspeed.ogx._provider import (
     DEFAULT_BASE_URL,
     OgxProvider,
 )
-from pydantic_ai_lightspeed.llamastack._transport import OgxServerTransport
+from pydantic_ai_lightspeed.ogx._transport import OgxServerTransport
 
 
 class TestOgxProviderProperties:
