+++ tests/unit/pydantic_ai_lightspeed/ogx/test_transport.py
@@ -1,4 +1,4 @@
-"""Unit tests for pydantic_ai_lightspeed.llamastack._transport module."""
+"""Unit tests for pydantic_ai_lightspeed.ogx._transport module."""
 
 # pylint: disable=protected-access
 
@@ -10,7 +10,7 @@ import httpx
 import pytest
 from pytest_mock import MockerFixture
 
-from pydantic_ai_lightspeed.llamastack._transport import (
+from pydantic_ai_lightspeed.ogx._transport import (
     OgxLibraryTransport,
     OgxServerTransport,
     _AsyncByteStream,
@@ -216,7 +216,7 @@ class TestHandleAsyncRequest:
 
         mock_func = mocker.AsyncMock(return_value={"id": "resp-1", "choices": []})
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
@@ -246,7 +246,7 @@ class TestHandleAsyncRequest:
 
         mock_func = mocker.AsyncMock(return_value=mock_stream_result())
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
@@ -265,7 +265,7 @@ class TestHandleAsyncRequest:
 
         mock_func = mocker.AsyncMock(return_value=[])
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
@@ -292,13 +292,13 @@ class TestHandleAsyncRequest:
 
         mock_func = mocker.AsyncMock(return_value={"id": "resp-1"})
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
 
         mock_ctx = mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.request_provider_data_context"
+            "pydantic_ai_lightspeed.ogx._transport.request_provider_data_context"
         )
 
         await transport.handle_async_request(request)
@@ -327,13 +327,13 @@ class TestHandleAsyncRequest:
 
         mock_func = mocker.AsyncMock(return_value={"id": "resp-1"})
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
 
         mock_ctx = mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.request_provider_data_context"
+            "pydantic_ai_lightspeed.ogx._transport.request_provider_data_context"
         )
 
         await transport.handle_async_request(request)
@@ -359,7 +359,7 @@ class TestHandleNonStreaming:
 
         mock_func = mocker.AsyncMock(return_value={"id": "model-1"})
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {"model_id": "test-model"}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
@@ -379,7 +379,7 @@ class TestHandleNonStreaming:
 
         mock_func = mocker.AsyncMock(return_value=None)
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
@@ -400,7 +400,7 @@ class TestHandleNonStreaming:
 
         mock_func = mocker.AsyncMock(return_value={"deleted": True})
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
@@ -433,7 +433,7 @@ class TestHandleStreaming:  # pylint: disable=too-few-public-methods
 
         mock_func = mocker.AsyncMock(return_value=mock_stream())
         mocker.patch(
-            "pydantic_ai_lightspeed.llamastack._transport.find_matching_route",
+            "pydantic_ai_lightspeed.ogx._transport.find_matching_route",
             return_value=(mock_func, {}, None, None),
         )
         transport._client._convert_body = mocker.Mock(side_effect=lambda f, b: b)
